+++ backend/tests/api/conftest.py
@@ -71,7 +71,7 @@ def recipe_yields():
 def recipe_time():
     return 30
 
-FIX_DATETIME = datetime(2023, 10, 1, 23, 59, 59)
+FIX_DATETIME = datetime(2025, 1, 1, 23, 59, 59)  # Wednesday
 
 @pytest.fixture
 def onboarded_client(client, admin_username, admin_name, admin_password):
@@ -217,26 +217,3 @@ def planned_recipe(user_client_with_household, household_id, recipe_with_items):
     
     return recipe_with_items  # Return recipe_id for convenience
 
-@pytest.fixture
-def planned_recipe_day_field_backwards_compatibility(user_client_with_household, household_id, recipe_with_items):
-    """Fixture that creates a meal plan with the test recipe"""
-    plan_data = {
-        'recipe_id': recipe_with_items,
-        "day": 0
-    }
-    response = user_client_with_household.post(
-        f'/api/household/{household_id}/planner/recipe',
-        json=plan_data
-    )
-    assert response.status_code == 200
-    
-    # Verify plan was created
-    response = user_client_with_household.get(
-        f'/api/household/{household_id}/planner'
-    )
-    assert response.status_code == 200
-    planned_meals = response.get_json()
-    assert len(planned_meals) > 0
-    assert any(meal['recipe']['id'] == recipe_with_items for meal in planned_meals)
-    
-    return recipe_with_items  # Return recipe_id for convenience
