+++ backend/tests/api/test_api_planner.py
@@ -29,24 +29,6 @@ def test_meal_planning_when_field(user_client_with_household, household_id, plan
     assert actual == expected
 
 
-def test_meal_planning_remove_by_day(user_client_with_household, household_id, planned_recipe_day_field_backwards_compatibility):
-    """Test removing meals from plan"""
-    # Remove from meal plan
-    response = user_client_with_household.delete(
-        f'/api/household/{household_id}/planner/recipe/{planned_recipe_day_field_backwards_compatibility}',
-        json={'day': 0}
-    )
-    assert response.status_code == 200
-
-    # Verify removal
-    response = user_client_with_household.get(
-        f'/api/household/{household_id}/planner'
-    )
-    assert response.status_code == 200
-    planned_meals = response.get_json()
-    assert not any(meal['recipe']['id'] == planned_recipe_day_field_backwards_compatibility for meal in planned_meals)
-
-
 def test_meal_planning_remove_by_datetime(user_client_with_household, household_id, planned_recipe):
     """Test removing meals from plan"""
     # Remove from meal plan
@@ -70,7 +52,7 @@ def test_recent_planned_recipes(user_client_with_household, household_id, planne
     # First remove the recipe from the plan
     response = user_client_with_household.delete(
         f'/api/household/{household_id}/planner/recipe/{planned_recipe}',
-        json={'day': 0}
+        json={'when': FIX_DATETIME.isoformat()}
     )
     assert response.status_code == 200
